\name{parspatstat-internal}
\title{Internal parspatstat objects}
\alias{owin.divide}
\alias{sample.ppp}
\alias{pool.K}
\alias{slave.parKest.ppp}
\alias{slave.parKest.parppp}
\alias{slave.parenvelope}
\alias{slave.parread.ppp}
\description{Internal parspatstat objects.}
\details{These are not to be called by the user.}
\keyword{internal}
