\name{parspatstat-package}
\alias{parspatstat-package}
\alias{parspatstat}
\docType{package}
\title{
Parallel spatial statistics
}
\description{
Provides parallel computing functionality (using Rmpi) to some methods in the spatstat package.
}
\details{
\tabular{ll}{
Package: \tab parspatstat\cr
Type: \tab Package\cr
Version: \tab 0.1-2\cr
Date: \tab 2012-09-21\cr
License: \tab GPL-2\cr
} 
Make sure the \code{Rmpi} package is installed correctly and slaves are spawned via \link{mpi.spawn.Rslaves} beforehand.
}
\author{
Jonathan Lee <jlee253@uwo.ca>

Maintainer: Jonathan Lee <jlee253@uwo.ca>
}

\keyword{ package }
%%\seealso{
%%}
%%\examples{

%%}
